+++ src/packages/documents/documents/tree/document-tree.repository.ts
@@ -1,7 +1,7 @@
 import { UMB_DOCUMENT_ROOT_ENTITY_TYPE } from '../entity.js';
 import { UmbDocumentTreeServerDataSource } from './document-tree.server.data-source.js';
 import type { UmbDocumentTreeItemModel, UmbDocumentTreeRootModel } from './types.js';
-import { UMB_DOCUMENT_TREE_STORE_CONTEXT } from './document-tree.store.js';
+import { UMB_DOCUMENT_TREE_STORE_CONTEXT } from './document-tree.store.context-token.js';
 import { UmbTreeRepositoryBase } from '@umbraco-cms/backoffice/tree';
 import type { UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import type { UmbApi } from '@umbraco-cms/backoffice/extension-api';
