+++ src/packages/documents/documents/tree/document-tree.store.context-token.ts
@@ -1,0 +1,4 @@
+import { UmbContextToken } from '@umbraco-cms/backoffice/context-api';
+import type { UmbDocumentTreeStore } from './document-tree.store.js';
+
+export const UMB_DOCUMENT_TREE_STORE_CONTEXT = new UmbContextToken<UmbDocumentTreeStore>('UmbDocumentTreeStore');
