+++ src/packages/documents/documents/tree/index.ts
@@ -4,6 +4,6 @@ export {
 	UMB_DOCUMENT_TREE_STORE_ALIAS,
 	UMB_DOCUMENT_TREE_ALIAS,
 } from './manifests.js';
-export { UMB_DOCUMENT_TREE_STORE_CONTEXT } from './document-tree.store.js';
+export { UMB_DOCUMENT_TREE_STORE_CONTEXT } from './document-tree.store.context-token.js';
 export type { UmbDocumentTreeStore } from './document-tree.store.js';
 export * from './types.js';
